% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_diffusion}
\alias{sim_diffusion}
\title{Simulate from a diffusion approximation to the Wright-Fisher model}
\usage{
sim_diffusion(s, p0, sigma, dx = 0.01, t_max, n_deme)
}
\arguments{
\item{s}{TODO.}

\item{p0}{TODO.}

\item{sigma}{TODO.}

\item{dx}{TODO.}

\item{t_max}{TODO.}

\item{n_deme}{TODO.}
}
\description{
Simulates from diffusion approximation to Wright-Fisher model.
}
