% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_posterior_prev}
\alias{get_posterior_prev}
\title{Posterior prevalence from forward-backward algorithm}
\usage{
get_posterior_prev(
  df_data,
  s,
  sigma,
  t_start = min(df_data$t),
  t_end = max(df_data$t),
  dt = 7,
  dx = 0.01,
  x_init = seq(0, 1, dx)
)
}
\arguments{
\item{df_data}{A data frame of observed data with columns: \code{pop} (population ID),
\code{t} (generation), \code{n_samp} (sample size), and \code{n_pos} (number of positive individuals).}

\item{s}{A scalar or vector of selection coefficients. If a vector, results are marginalized over the values.}

\item{sigma}{A scalar or vector of diffusion parameters, corresponding to the values of \code{s}.}

\item{t_start}{The first date to include in the trajectory (default is min of data).}

\item{t_end}{The last date to include in the trajectory (default is max of data).}

\item{dt}{The time step (in days) between prediction dates (default is 7 days).}

\item{dx}{Step size for discretizing allele frequencies (default is \code{0.01}).}

\item{x_init}{A vector specifying the initial allele frequency distribution (default is \code{seq(0, 1, dx)}).}
}
\value{
A data frame with columns:
\describe{
  \item{pop}{Population ID.}
  \item{p}{Allele frequency bin (from \code{x_init}).}
  \item{t}{Date.}
  \item{post}{Posterior probability of allele frequency \code{p} at time \code{t}.}
}
}
\description{
Runs the forward-backward algorithm to estimate the posterior distribution of allele prevalence
over time, given observed data and model parameters. Supports multiple populations and can marginalize over
uncertainty in parameters by averaging over a set of posterior samples.
}
