% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sample_prev}
\alias{sample_prev}
\title{Sample from a simulated prevalence trajectory}
\usage{
sample_prev(df_sim, df_sample)
}
\arguments{
\item{df_sim}{TODO.}

\item{df_sample}{TODO.}
}
\description{
TODO
}
