% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{sim_WF}
\alias{sim_WF}
\title{Simulate from a Wright-Fisher model}
\usage{
sim_WF(s, p0, N, t_max, n_deme)
}
\arguments{
\item{s}{TODO.}

\item{p0}{TODO.}

\item{N}{TODO.}

\item{t_max}{TODO.}

\item{n_deme}{TODO.}
}
\description{
Simulates from Wright-Fisher model with drift and selection.
}
